+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 - 2023-02-01
+- **Other Features**
+  - Added `Spiral\PhoneNumber\Validator\Checker\PhoneNumberChecker` for the
+    [Spiral Validator](https://github.com/spiral/validator) package.
+
 ## 1.0.0 - 2023-01-08
 
-- initial release
+ - initial release